==> 1_Creation_de_3_dataset.ipynb <==
# Projet Long : étude de champs de linéaires 

#import des packages requis
import numpy as np 
import pandas as pd 
import seaborn as sns
import matplotlib.pyplot as plt

# Import des données : champ B pour Bretagne

champB = pd.read_csv('SummarySegment-B.csv', delimiter=' ')
champB

On drop landscape car on a toujours B (qui signifie Bretagne)

champB = champB.drop('Landscape', axis = 1)

On créer ensuite 3 dataframe : un pour les haies, un pour les canaux, un pour les routes :


haies = champB[champB['type'] == 'H']
haies = haies.drop_duplicates(['id', 'Lng', 'typeH','Ang','X1','X2','Y1','Y2'])


canaux = champB[champB['type'] == 'C']
print('avant modif =',canaux.shape)
canaux = canaux.drop_duplicates(['id', 'Lng', 'typeH','Ang','X1','X2','Y1','Y2'])
print('après modif =',canaux.shape)

routes = champB[champB['type'] == 'R']
print('avant modif =',routes.shape)
routes = routes.drop_duplicates(['id', 'Lng', 'typeH','Ang','X1','X2','Y1','Y2'])
print('avant modif =',routes.shape)

On supprime les doublons et on regarde la taille du dataset avant et après cette opération :

print('avant modif =',haies.shape)
haies = haies.drop_duplicates(['id', 'Lng', 'typeH','Ang','X1','X2','Y1','Y2'])
print('après modif =',haies.shape)

On les enregistre pour les avoir au format csv

haies.to_csv('1_haies.csv')
canaux.to_csv('1_canaux.csv')
routes.to_csv('1_routes.csv')
