==> 5_Rajout_de_variables.ipynb <==
#import des packages requis
import numpy as np 
import pandas as pd 
import seaborn as sns 
import matplotlib.pyplot as plt 

def classification_haies(haies_df, canaux_df, routes_df, distance_max_canal, distance_max_route):
    '''Prend en entrée un dataframe de haies, et rajoute 3 colonnes : 
    - une pour les haies proches des canaux (0 : pas proche , 1 : proche)
    - une pour les haies proches des routes (0 : pas proche , 1 : proche)
    - une pour les haies perpandiculaires aux routes ((0 : pas perpandiculaire , 1 : perpandiculaire))
    '''

    haies = haies_df.copy()
    canaux = canaux_df.copy()
    routes = routes_df.copy()

    # On créer nos 3 nouvelles colonnes, remplie de 0
    haies['proche route'] = 0
    haies['proche canal'] = 0
    haies['perp route'] = 0
    haies['perp canal'] = 0
    haies['paral route'] = 0
    haies['paral canal'] = 0

    # On parcourt toutes les haies du dataset
    for i, haie_courante in haies.iterrows():
        #On stocke dans r et c les valeurs de row et col de la haie courante
        r = haie_courante['row']
        c = haie_courante['col']
        #On stocke dans canaux_proches et routes_proches .... ça parle de soit même !
        canaux_proches = canaux[(abs(canaux['row'] - r) <= 1) & (abs(canaux['col'] - c) <= 1)]
        routes_proches = routes[(abs(routes['row'] - r) <= 1) & (abs(routes['col'] - c) <= 1)]

        #On parcourt les routes proches de la haie courante
        for _, route_courant in routes_proches.iterrows():
            distance = np.sqrt(abs(haie_courante['XG'] - route_courant['XG'])**2 + abs(haie_courante['YG'] - route_courant['YG'])**2)
            diff_angle = abs(route_courant['Ang'] - haie_courante['Ang'])
            if distance < distance_max_route:
                haies.at[i, 'proche route'] = 1
                if diff_angle < 2.35 and diff_angle > 0.785 : 
                    haies.at[i, 'perp route'] = 1
                elif diff_angle < 0.78 :
                    haies.at[i,'paral route'] = 1
                break

        #On parcourt les canaux proches de la haie courante 
        for _, canal_courant in canaux_proches.iterrows():
            distance = np.sqrt(abs(haie_courante['XG'] - canal_courant['XG'])**2 + abs(haie_courante['YG'] - canal_courant['YG'])**2)
            diff_angle = abs(canal_courant['Ang'] - haie_courante['Ang'])
            if distance < distance_max_canal:
                haies.at[i, 'proche canal'] = 1
                if diff_angle < 2.35 and diff_angle > 0.785 : 
                    haies.at[i, 'perp canal'] = 1
                elif diff_angle < 0.78 :
                    haies.at[i,'paral canal'] = 1
                break

    return haies


distance_max_canal = 30
distance_max_route = 40
haie_fusion = pd.read_csv('3_haie_fusion.csv', delimiter=',',index_col=0)
canaux = pd.read_csv('1_canaux.csv', delimiter=',',index_col=0)
routes = pd.read_csv('1_routes.csv', delimiter=',',index_col=0)
haies_classfiees_sans_fusion = classification_haies(haie_fusion, canaux, routes, distance_max_canal, distance_max_route)
haies_classfiees_sans_fusion.to_csv("5_sf_haies_classfiees.csv")

haies_proche_canal_sf = haies_classfiees_sans_fusion[haies_classfiees_sans_fusion['proche canal'] == 1]
haies_proche_route_sf = haies_classfiees_sans_fusion[haies_classfiees_sans_fusion['proche route'] == 1]
haies_perp_canal_sf = haies_classfiees_sans_fusion[haies_classfiees_sans_fusion['perp canal'] == 1]
haies_perp_route_sf = haies_classfiees_sans_fusion[haies_classfiees_sans_fusion['perp route'] == 1]
haies_paral_canal_sf = haies_classfiees_sans_fusion[haies_classfiees_sans_fusion['paral canal'] == 1]
haies_paral_route_sf = haies_classfiees_sans_fusion[haies_classfiees_sans_fusion['paral route'] == 1]

haies_proche_canal_sf.to_csv("5_sf_haies_proche_canal.csv")
haies_proche_route_sf.to_csv("5_sf_haies_proche_route.csv")
haies_perp_canal_sf.to_csv("5_sf_haies_perp_canal.csv")
haies_perp_route_sf.to_csv("5_sf_haies_perp_route.csv")
haies_paral_canal_sf.to_csv("5_sf_haies_paral_canal.csv")
haies_paral_route_sf.to_csv("5_sf_haies_paral_route.csv")

distance_max_canal = 30
distance_max_route = 40
haie_fusion = pd.read_csv('3_haie_fusion.csv', delimiter=',',index_col=0)
canaux_fusion = pd.read_csv('3_canaux_fusion.csv', delimiter=',',index_col=0)
routes_fusion = pd.read_csv('3_routes_fusion.csv', delimiter=',',index_col=0)
haies_classfiees_avec_fusion = classification_haies(haie_fusion, canaux_fusion, routes_fusion, distance_max_canal, distance_max_route)
haies_classfiees_avec_fusion.to_csv("7_af_haies_classfiees.csv")

haies_proche_canal_af = haies_classfiees_avec_fusion[haies_classfiees_avec_fusion['proche canal'] == 1]
haies_proche_route_af = haies_classfiees_avec_fusion[haies_classfiees_avec_fusion['proche route'] == 1]
haies_perp_canal_af = haies_classfiees_avec_fusion[haies_classfiees_avec_fusion['perp canal'] == 1]
haies_perp_route_af = haies_classfiees_avec_fusion[haies_classfiees_avec_fusion['perp route'] == 1]
haies_paral_canal_af = haies_classfiees_avec_fusion[haies_classfiees_avec_fusion['paral canal'] == 1]
haies_paral_route_af = haies_classfiees_avec_fusion[haies_classfiees_avec_fusion['paral route'] == 1]

haies_proche_canal_af.to_csv("7_af_haies_proche_canal.csv")
haies_proche_route_af.to_csv("7_af_haies_proche_route.csv")
haies_perp_canal_af.to_csv("7_af_haies_perp_canal.csv")
haies_perp_route_af.to_csv("7_af_haies_perp_route.csv")
haies_paral_canal_af.to_csv("7_af_haies_paral_canal.csv")
haies_paral_route_af.to_csv("7_af_haies_paral_route.csv")

# classification en fonction de la nature du sol

haies_classfiees = pd.read_csv('5_sf_haies_classfiees.csv')

ville = pd.read_csv('haie_ville.csv')
pastures = pd.read_csv('haie_pastures.csv')
complex_cult = pd.read_csv('haie_complex_cultivation_pattern.csv')
natural_vegetation = pd.read_csv('haie_agriculture_with_natural_vegetation.csv')
non_irrigated = pd.read_csv('haies_non_irrigated.csv')

# codage pour l'appartenance des sols : 0 si la haie n'appartient pas au sol considéré, 1 sinon


haies_classfiees['ville'] = 0
haies_classfiees['pastures'] = 0
haies_classfiees['complex_cult'] = 0
haies_classfiees['natural_vegetation'] = 0
haies_classfiees['non_irrigated'] = 0


# on parcourt ici toutes nos, et dès lors que l'on tombe sur une haie qui appartient à l'un des 5 datasets des sols, 
# on met à jour la variable d'appartenance du sol correspondante
for i,haie_courante in haies_classfiees.iterrows():
  id_courant = haie_courante['id']

  if id_courant in ville['id'].values :
    haies_classfiees.at[i,'ville'] = 1

  if id_courant in pastures['id'].values :
    haies_classfiees.at[i,'pastures'] = 1

  if id_courant in complex_cult['id'].values :
    haies_classfiees.at[i,'complex_cult'] = 1

  if id_courant in natural_vegetation['id'].values :
    haies_classfiees.at[i,'natural_vegetation'] = 1

  if id_courant in non_irrigated['id'].values :
    haies_classfiees.at[i,'non_irrigated'] = 1

haies_classfiees.to_csv('haies_classfiees.csv')
